oss-scrape/requirements.txt
beautifulsoup4==4.7.1
requests==2.21.0
python-dateutil==2.8.0
python-dotenv==0.10.1
influxdb==5.2.1